+++ app/api/ada/lib/storage/bridge/delegationUtils.js
@@ -18,10 +18,10 @@ import type {
 } from '../models/PublicDeriver/interfaces';
 import {
   filterAddressesByStakingKey,
+  delegationTypeToResponse,
 } from './utils';
 import type {
   AccountStateDelegation,
-  PoolTuples,
   AccountStateSuccess,
 } from '../../state-fetch/types';
 import { TxStatusCodes } from '../database/primitives/enums';
@@ -197,39 +197,5 @@ export function certificateToPoolList(
     default: {
       throw new Error(`${nameof(certificateToPoolList)} unexpected certificate kind ${kind}`);
     }
-  }
-}
-
-export function delegationTypeToResponse(
-  type: RustModule.WalletV3.DelegationType,
-): AccountStateDelegation {
-  const kind = type.get_kind();
-  switch (kind) {
-    case RustModule.WalletV3.DelegationKind.NonDelegated: return { pools: [], };
-    case RustModule.WalletV3.DelegationKind.Full: {
-      const poolId = type.get_full();
-      if (poolId == null) {
-        throw new Error(`${nameof(delegationTypeToResponse)} Should never happen`);
-      }
-      return {
-        pools: [[poolId.to_string(), 1]]
-      };
-    }
-    case RustModule.WalletV3.DelegationKind.Ratio: {
-      const ratios = type.get_ratios();
-      if (ratios == null) {
-        throw new Error(`${nameof(delegationTypeToResponse)} Should never happen`);
-      }
-      const poolTuples: Array<PoolTuples> = [];
-      const pools = ratios.pools();
-      for (let i = 0; i < pools.size(); i++) {
-        const pool = pools.get(i);
-        poolTuples.push([pool.pool().to_string(), pool.parts()]);
-      }
-      return {
-        pools: poolTuples,
-      };
-    }
-    default: throw new Error(`${nameof(delegationTypeToResponse)} unexpected kind ${kind}`);
   }
 }
